CODY-92 move fixtures to conftest.py
